+++ src/core/functions/distribution/ProbabilityDensityFunction.h
@@ -131,7 +131,7 @@ void RevBayesCore::ProbabilityDensityFunction<valueType>::update( void )
     
     if ( useLog == true )
     {
-        *value = dist->computeLnProbability();
+        *value = (double)dist->computeLnProbability();
     }
     else
     {
